Add spec for java installation
